+++ openshift-client/src/main/java/io/fabric8/openshift/client/DefaultOpenShiftClient.java
@@ -89,6 +89,7 @@ import io.fabric8.kubernetes.client.dsl.BatchAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.CertificatesAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.Createable;
 import io.fabric8.kubernetes.client.dsl.DiscoveryAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.EventingAPIGroupDSL;
 import io.fabric8.kubernetes.client.dsl.FunctionCallable;
 import io.fabric8.kubernetes.client.dsl.InOutCreateable;
 import io.fabric8.kubernetes.client.dsl.KubernetesListMixedOperation;
@@ -477,6 +478,11 @@ public class DefaultOpenShiftClient extends BaseClient implements NamespacedOpen
   @Override
   public DiscoveryAPIGroupDSL discovery() {
     return delegate.discovery();
+  }
+
+  @Override
+  public EventingAPIGroupDSL events() {
+    return delegate.events();
   }
 
   @Override